Fix the bug in triangle-counting/playground/output_remove_duplicate.cpp, add support for large graphs, pre-processing not correct yet

+++ triangle-counting/playground/output_remove_duplicate.cpp
@@ -131,7 +131,7 @@ int main(int argc, char *argv[]) {
         auto *relative_off = (uint32_t *) malloc(sizeof(uint32_t) * num_edges);
 #pragma omp parallel num_threads(max_omp_threads)
         {
-            SelectNotFOMP(histogram, edge_lst_buffer, edge_lst, relative_off, num_edges, [edge_lst](uint32_t it) {
+            SelectNotFOMP(histogram, edge_lst_buffer, edge_lst, relative_off, num_edges, [edge_lst](size_t it) {
                 return edge_lst[it].first == edge_lst[it].second || (it > 0 && edge_lst[it - 1] == edge_lst[it]);
             });
         }
